lambda/requirements.txt
matplotlib
numpy
pillow
scipy
tensorflow==2.4.1
